Updated intake. added ColorMatcher to controlPanel incase we use that

+++ src/main/java/frc/team2412/robot/Subsystems/constants/IntakeConstants.java
@@ -3,9 +3,9 @@ package frc.team2412.robot.Subsystems.constants;
 import edu.wpi.first.wpilibj.DoubleSolenoid;
 
 public class IntakeConstants {
-	
-	public final static int MAX_INTAKE_SPEED = 1;
-	
+
+	public final static double MAX_INTAKE_SPEED = 1;
+
 	public static enum IntakeState {
 
 		// Creates a value called up which is equal to kForward
@@ -21,14 +21,14 @@ public class IntakeConstants {
 			this.value = value;
 		}
 	}
-	
+
 	public static enum IntakeLastMotor {
 		FRONT(1), BOTH(0), BACK(-1);
-		
-		public int value;
-		
-		private IntakeLastMotor(int value) {
-			this.value = value;			
+
+		public double value;
+
+		private IntakeLastMotor(double value) {
+			this.value = value;
 		}
 	}
